+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-9157 Added support for generating file-related URLs on embedded space automation view
 * VFS-11969 Added providers offline info on data tile on space overview
 * VFS-12006 Fixed missing attributes in file/group data spec editor during lambda creation
 * VFS-11760 Fixed global shares sidebar crashing when user has no view privilege for some space